renderers: add a completeInitialization callback
Gives a chance to perform initialization work once we are sure the renderer
has been initialized with all the default values.

This is needed to perform a viewAll with the SerenityRenderer which is an
immediate operation upon model loading (so takes camera properties at the time
the aspects are first created) unlike the Qt3DRenderer where viewAll is
delayed by a couple of frames.

Change-Id: I84dbcd2042aae926e5a22dcaf9b0c12721992aa0
Reviewed-on: https://codereview.kdab.com/c/schneider-digital/allegiance/+/150966
Reviewed-by: Mike Krus <mike.krus@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>