Serenity: convert diffuseTexture from Linear to sRGB
We assume that the provided textures are exported with an sRGB color
profile. Yet since we load texture as Linear RGB, we need to convert
from Linear to sRGB in the shaders.

We don't need to bother with converting the final output of our shaders
to sRGB since Serenity selects an sRGB swapchain (so this conversion is
implicitly done for us in hardware).

Change-Id: Ie6ba6ae8022863cd3ad1f27b825a685ad41c888c
Reviewed-on: https://codereview.kdab.com/c/schneider-digital/allegiance/+/151784
Reviewed-by: Mike Krus <mike.krus@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>