Fix viewAll not working following model being loaded
Turns out that they was a conflict where the mouse being moved
in the file dialog to select the new file was leading to the camera
being rotated around right after the scene got loaded (since we have
a global event filter).
Changing the rotate condition to ensure we have been correctly pressed
beforehand solves it.

Task-Id: SDSDR-259
Change-Id: I3ab16b424b927e84f0bf697aed3272104d7b06fa
Reviewed-on: https://codereview.kdab.com/c/schneider-digital/allegiance/+/151798
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Mike Krus <mike.krus@kdab.com>